move play store to app dir
