+++ apps/client/src/play/hooks/useSetAvatar.ts
@@ -1,6 +1,6 @@
 import { getTempUpload } from "../../../app/api/temp/helper";
+import { usePlayStore } from "../../../app/play/[id]/store";
 import { env } from "../../env/client.mjs";
-import { usePlayStore } from "../../play/store";
 import { LocalStorageKey } from "../constants";
 import { sendToHost } from "./useHost";
